Move System.Speech p/invokes into common directory
