+++ src/libraries/Common/src/Interop/Windows/WinMm/Interop.waveOutPause.cs
@@ -1,0 +1,21 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Runtime.InteropServices;
+
+internal static partial class Interop
+{
+    internal static partial class WinMM
+    {
+        /// <summary>
+        /// This function pauses playback on a specified waveform output device. The
+        /// current playback position is saved. Use waveOutRestart to resume playback
+        /// from the current playback position.
+        /// </summary>
+        /// <param name="hwo">Handle to the waveform-audio output device.</param>
+        /// <returns>MMSYSERR</returns>
+        [DllImport(Libraries.WinMM)]
+        internal static extern MMSYSERR waveOutPause(IntPtr hwo);
+    }
+}
