+++ src/libraries/Common/src/Interop/Windows/WinMm/Interop.waveOutReset.cs
@@ -1,0 +1,21 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Runtime.InteropServices;
+
+internal static partial class Interop
+{
+    internal static partial class WinMM
+    {
+        /// <summary>
+        /// This function stops playback on a specified waveform output device and
+        /// resets the current position to 0. All pending playback buffers are marked
+        /// as done and returned to the application.
+        /// </summary>
+        /// <param name="hwo">Handle to the waveform-audio output device.</param>
+        /// <returns>MMSYSERR</returns>
+        [DllImport(Libraries.WinMM)]
+        internal static extern MMSYSERR waveOutReset(IntPtr hwo);
+    }
+}
